pyproject.toml
[project]
name = "manga-translator-cli"
version = "0.1.0"
description = "A CLI tool for automatically translating manga pages from Japanese to English. Detects speech bubbles, extracts Japanese text using OCR, translates to English, and renders the translated text back onto images with proper alignment."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.0",
    "protobuf>=3.20.0",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.5",
    "ultralytics>=8.0.0",
    "huggingface-hub>=0.20.0",
    "ctranslate2>=4.6.3",
    "opencv-python>=4.8.0",
    "einops>=0.8.1",
    "accelerate>=1.12.0",
    "rich>=13.0.0",
]


[project.scripts]
manga-translate = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
index-strategy = "unsafe-best-match"

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
torchvision = { index = "pytorch-cu128" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
